backend/app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.9.13
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
